requirements.txt
requests
pandas
python-dotenv
google-cloud-bigquery
google-api-core
pyarrow
pandas-gbq
